INS-2588: bootstrap codebase moved to gateway
