+++ configuration/servicenetwork.go
@@ -18,7 +18,7 @@ package configuration
 
 // ServiceNetwork is configuration for ServiceNetwork.
 type ServiceNetwork struct {
-	Skip             int // magic number that indicates what delta after last ignored pulse we should wait
+	// Skip             int // magic number that indicates what delta after last ignored pulse we should wait
 	CacheDirectory   string
 	Consensus        Consensus
 	ConsensusEnabled bool
@@ -35,7 +35,7 @@ type Consensus struct {
 // NewServiceNetwork creates a new ServiceNetwork configuration.
 func NewServiceNetwork() ServiceNetwork {
 	return ServiceNetwork{
-		Skip:             10,
+		// Skip:             10,
 		CacheDirectory:   "network_cache",
 		Consensus:        NewConsensus(),
 		ConsensusEnabled: true,
